move class to separate file
